# PERF: improve perf of float-based timeseries plotting