--- a/doc/source/whatsnew/v0.20.0.txt
+++ b/doc/source/whatsnew/v0.20.0.txt
@@ -279,7 +279,8 @@ Performance Improvements
 
 - Improved performance of ``pd.wide_to_long()`` (:issue:`14779`)
 - Increased performance of ``pd.factorize()`` by releasing the GIL with ``object`` dtype when inferred as strings (:issue:`14859`)
-
+- Improved performance of timeseries plotting with an irregular DatetimeIndex
+  (or with ``compat_x=True``) (:issue:`15073`).
 
 
 .. _whatsnew_0200.bug_fixes: