@jreback 2017-01-10
https://travis-ci.org/pandas-dev/pandas/jobs/190356982

I think this broke something on the slow 3.4 build ?